+++ project2/Database/src/main/java/ru/cinimex/learn/dao/ICustomerDao.java
@@ -14,5 +14,5 @@ import java.util.List;
 
     public void insert( Customer customer);
 
-    public Customer getByPhone(Integer phone);
+    public Customer getByPhone(Long phone);
 }
